requirements.txt
selenium==4.23.1
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
tqdm==4.66.4
pandas==2.2.2
python-dotenv==1.0.1
